Eliminate the sets
It's more code for not too much gain since we're only dealing with tiny
slices and not really looking through them that often.
